Initial commit of refactoring "view achieved goal" into "view goal details"
