+++ server/routes/URLs.ts
@@ -4,6 +4,7 @@ export default {
   ADD_STEPS: '/goal/:uuid/add-steps',
   PLAN_OVERVIEW: '/plan',
   VIEW_ACHIEVED_GOAL: '/view-achieved-goal/:uuid',
+  VIEW_REMOVED_GOAL: '/view-removed-goal/:uuid',
   GOALS: '/goals',
   GOALS_ORDER: '/goals/:type/:uuid/:operation',
   REMOVE_GOAL: '/remove-goal/:uuid',
