--- server/routes/viewAchievedGoal/locale.json
@@ -1,12 +1,0 @@
-{
-  "en": {
-    "page": {
-      "title": "View achieved goal - {{ applicationName }}",
-      "errorTitle": "Error: View achieved goal - {{ applicationName }}"
-    },
-    "mainHeading": {
-      "title": "Goal details"
-    },
-    "markedAsAchieved": "Marked as achieved on {{ achievedOn }}."
-  }
-}
