+++ server/routes/viewGoalDetails/routes.ts
@@ -1,9 +1,10 @@
 import { Router } from 'express'
 import URLs from '../URLs'
-import ViewAchievedGoalController from './ViewAchievedGoalController'
+import ViewGoalDetailsController from './ViewGoalDetailsController'
 
 export default function setupViewAchievedGoalRoutes(router: Router) {
-  const controller = new ViewAchievedGoalController()
+  const controller = new ViewGoalDetailsController()
 
   router.get(URLs.VIEW_ACHIEVED_GOAL, controller.get)
+  router.get(URLs.VIEW_REMOVED_GOAL, controller.get)
 }
