+++ server/views/pages/plan.njk
@@ -235,7 +235,7 @@
                         {% elseif goal.status === 'REMOVED' %}
                             {% set actions = [
                                 {
-                                    href: '#',
+                                    href: '/view-removed-goal/' + goal.uuid,
                                     text: locale.goalSummaryCard.actions.removed.viewDetails
                                 },
                                 {
