+++ server/views/pages/view-goal-details.njk
@@ -3,7 +3,8 @@
 {% extends "../partials/layout.njk" %}
 
 {% set locale = interpolate(locale, {
-    achievedOn: data.goal.statusDate | formatSimpleDate
+    statusChangedDate: data.goal.statusDate | formatSimpleDate,
+    status: data.goal.status | lower
 }) %}
 
 {% block pageTitle %}
@@ -25,9 +26,13 @@
                     (and {% for aon in data.goal.relatedAreasOfNeed %}{{ aon.name | lower }}{% if loop.index < data.goal.relatedAreasOfNeed.length %}, {% endif %}{% endfor %})
                 {% endif %}
             </span>
-            <h1 class="govuk-heading-l govuk-!-margin-bottom-2">{{ data.goal.title }}</h1>
-            <p class="govuk-body">{{ locale.markedAsAchieved }}</p>
-
+            <h1 class="govuk-heading-l govuk-!-margin-bottom-2">{{ locale.mainHeading.title }}</h1>
+            <h2 class="govuk-heading-m govuk-!-margin-bottom-2">{{ locale.goalTitlePrefix }} {{ data.goal.title }}</h2>
+            {% if data.goal.status == 'ACHIEVED' %}
+                <p class="govuk-body">{{ locale.markedAsAchieved }}</p>
+            {% elif data.goal.status == 'REMOVED' %}
+                <p class="govuk-body">{{ locale.markedAsRemoved }}</p>
+            {% endif %}
             <table class="govuk-table goal-summary-card__steps">
                 <thead class="govuk-table__head">
                 <tr class="govuk-table__row">
